added fake auth and router to connect two pages
