+++ template.json
@@ -5,11 +5,12 @@
       "@atlaskit/css-reset": "6.0.5",
       "@atlaskit/form": "8.2.2",
       "@atlaskit/menu": "^0.7.4",
+      "@atlaskit/icon": "21.4.0",
       "@atlaskit/side-navigation": "^0.8.3",
       "@atlaskit/button": "15.1.5",
       "@atlaskit/checkbox": "12.0.1",
       "@atlaskit/textfield": "5.0.1",
-      "@atlaskit/toggle": "12.0.4",
+      "@atlaskit/theme": "11.0.2",
       "@testing-library/jest-dom": "5.11.10",
       "@testing-library/react": "11.2.5",
       "@testing-library/user-event": "13.0.16",
@@ -17,6 +18,8 @@
       "@types/react": "17.0.3",
       "@types/react-dom": "17.0.3",
       "@types/jest": "26.0.22",
+      "@types/react-router-dom": "^5.1.7",
+      "react-router-dom": "^5.2.0",
       "typescript": "4.2.3",
       "styled-components": "5.2.1"
     },
