+++ template/src/app.tsx
@@ -1,25 +1,16 @@
+import Home from './home';
 import Login from './login';
-import homepage from './hero-illo.svg';
 
-const appStyles = {
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'space-evenly',
-  backgroundColor: '#fefefe',
-};
-
-const heroImageStyles = {
-  maxWidth: '36%',
-  opacity: 1,
-};
+import { Switch, Route } from 'react-router-dom';
 
 const App = () => {
   return (
-    <div style={appStyles}>
-      <img src={homepage} style={heroImageStyles} alt="" />
-      <Login />
-    </div>
+    <Switch>
+      <Route path="/login" component={Login}  />
+      <Route path="/" component={Home} />
+    </Switch>
   )
 };
 
 export default App;
+
