+++ template/src/home/sidenav/next-gen-project-icon.tsx
