+++ template/src/home/sidenav/sample-footer.tsx
