+++ template/src/home/sidenav/sample-logo.tsx
